Close restore provenance TOCTOU gaps
