+++ packages/backup-control-plane/wrangler.jsonc
@@ -27,7 +27,6 @@
 		"ENABLE_PRODUCTION_D1_BACKUPS": "false",
 		"BACKUP_BENCHMARK_APPROVED": "false",
 		"SOURCE_ACCOUNT_ID": "REPLACE_WITH_ACCOUNT_UUID",
-		"SOURCE_ACCOUNT_NAME": "REPLACE_WITH_ACCOUNT_NAME",
 		"SOURCE_DATABASE_ID": "REPLACE_WITH_DATABASE_UUID",
 		"SOURCE_DATABASE_NAME": "REPLACE_WITH_DATABASE_NAME",
 		"ALLOWED_SOURCE_ACCOUNT_IDS": "REPLACE_WITH_ACCOUNT_UUID",
